requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.2
equinox==0.12.2
filelock==3.18.0
fsspec==2025.5.1
idna==3.10
jax==0.6.1
jaxlib==0.6.1
jaxtyping==0.3.2
Jinja2==3.1.6
MarkupSafe==3.0.2
ml_dtypes==0.5.1
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
opt_einsum==3.4.0
pillow==11.2.1
requests==2.32.3
scipy==1.15.3
sympy==1.14.0
torch==2.7.0
torchvision==0.22.0
triton==3.3.0
typing_extensions==4.13.2
urllib3==2.4.0
wadler_lindig==0.1.6
